pyproject.toml
[tool.poetry]
name = "drf-template"
version = "0.1.0"
description = "Python service"
authors = []
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.6"
python-decouple = "^3.8"
django-cors-headers = "^4.1.0"
pre-commit = "^3.3.3"
coreapi = "^2.3.3"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
